Add files via upload
This is a more difficult part than previous three. That is why you nee to focus on hardly on every steps of tasks. This is my way of decision, maybe it will be better to use totally vectorization for cost function(IMHO).
Later, I will try to do it with vectorization(without for-loop).